[effects/showfps] Calculate "correct" FPS
Summary:
cca0e15b455ba6 broke code that calculates the value of the fps counter.

The frames field stores timestamps of previous frames, not time between
each two consecutive frames.

This change doesn't attempt to make the show fps compute more accurate
performance metrics, e.g. how much times it takes to execute rendering
commands on the GPU, etc.

Test Plan: The Show FPS effect displays more sensible values.

Reviewers: #kwin, davidedmundson

Reviewed By: #kwin, davidedmundson

Subscribers: davidedmundson, broulik, kwin

Tags: #kwin

Differential Revision: https://phabricator.kde.org/D29247

+++ effects/showfps/showfps.h
@@ -89,7 +89,7 @@ private:
     int paint_size[ NUM_PAINTS ]; // number of pixels painted
     int paints_pos;  // position in the queue
     enum { MAX_FPS = 200 };
-    int frames[ MAX_FPS ]; // (sec*1000+msec) of the time the frame was done
+    qint64 frames[ MAX_FPS ]; // the time when the frame was done
     int frames_pos; // position in the queue
     double alpha;
     int x;
